chore: Move id.rs into mod.rs
